// src/gen/ride/wallet/v1alpha1/wallet_service.ts
export interface CreateAccountRequest {
    /**
     * @generated from protobuf field: string uid = 1;
     */
    uid: string;
}
/**
 * @generated from protobuf message ride.wallet.v1alpha1.CreateAccountResponse
 */
export interface CreateAccountResponse {
    /**
     * @generated from protobuf field: ride.wallet.v1alpha1.Account account = 1;
     */
    account?: Account;
}
/**
 * @generated from protobuf message ride.wallet.v1alpha1.GetAccountRequest
 */
export interface GetAccountRequest {
    /**
     * @generated from protobuf field: string account_id = 1;
     */
    accountId: string;
    /**
     * @generated from protobuf field: optional google.protobuf.FieldMask field_mask = 99;
     */
    fieldMask?: FieldMask;
}
/**
 * @generated from protobuf message ride.wallet.v1alpha1.GetAccountResponse
 */
export interface GetAccountResponse {
    /**
     * @generated from protobuf field: ride.wallet.v1alpha1.Account account = 1;
     */
    account?: Account;
}
/**
 * @generated from protobuf message ride.wallet.v1alpha1.GetAccountByUidRequest
 */
export interface GetAccountByUidRequest {
    /**
     * @generated from protobuf field: string uid = 1;
     */
    uid: string;
    /**
     * @generated from protobuf field: optional google.protobuf.FieldMask field_mask = 99;
     */
    fieldMask?: FieldMask;
}
/**
 * @generated from protobuf message ride.wallet.v1alpha1.GetAccountByUidResponse
 */
export interface GetAccountByUidResponse {
    /**
     * @generated from protobuf field: ride.wallet.v1alpha1.Account account = 1;
     */
    account?: Account;
}
/**
 * @generated from protobuf message ride.wallet.v1alpha1.Account
 */
export interface Account {
    /**
     * @generated from protobuf field: string account_id = 1;
     */
    accountId: string;
    /**
     * @generated from protobuf field: string uid = 2;
     */
    uid: string;
    /**
     * @generated from protobuf field: int32 balance = 3;
     */
    balance: number;
    /**
     * @generated from protobuf field: google.protobuf.Timestamp create_time = 7;
     */
    createTime?: Timestamp;
    /**
     * @generated from protobuf field: google.protobuf.Timestamp update_time = 8;
     */
    updateTime?: Timestamp;
}
/**
 * @generated from protobuf message ride.wallet.v1alpha1.CreateTransactionsRequest
 */
export interface CreateTransactionsRequest {
    /**
     * @generated from protobuf field: repeated ride.wallet.v1alpha1.CreateTransactionsRequest.Entry transactions = 2;
     */
    transactions: CreateTransactionsRequest_Entry[];
}
/**
 * @generated from protobuf message ride.wallet.v1alpha1.CreateTransactionsRequest.Entry
 */
export interface CreateTransactionsRequest_Entry {
    /**
     * @generated from protobuf field: string account_id = 2;
     */
    accountId: string;
    /**
     * @generated from protobuf field: int32 amount = 3;
     */
    amount: number;
    /**
     * @generated from protobuf field: ride.wallet.v1alpha1.TransactionType type = 4;
     */
    type: TransactionType;
}
/**
 * @generated from protobuf message ride.wallet.v1alpha1.CreateTransactionsResponse
 */
export interface CreateTransactionsResponse {
    /**
     * @generated from protobuf field: string batch_id = 1;
     */
    batchId: string;
    /**
     * @generated from protobuf field: repeated string transaction_ids = 2;
     */
    transactionIds: string[];
}
/**
 * @generated from protobuf message ride.wallet.v1alpha1.GetTransactionRequest
 */
export interface GetTransactionRequest {
    /**
     * @generated from protobuf field: string transaction_id = 1;
     */
    transactionId: string;
    /**
     * @generated from protobuf field: optional google.protobuf.FieldMask field_mask = 99;
     */
    fieldMask?: FieldMask;
}
/**
 * @generated from protobuf message ride.wallet.v1alpha1.GetTransactionResponse
 */
export interface GetTransactionResponse {
    /**
     * @generated from protobuf field: ride.wallet.v1alpha1.Transaction transaction = 1;
     */
    transaction?: Transaction;
}
/**
 * @generated from protobuf message ride.wallet.v1alpha1.ListTransactionsByBatchIdRequest
 */
export interface ListTransactionsByBatchIdRequest {
    /**
     * @generated from protobuf field: string batch_id = 1;
     */
    batchId: string;
    /**
     * @generated from protobuf field: optional google.protobuf.FieldMask field_mask = 99;
     */
    fieldMask?: FieldMask;
}
/**
 * @generated from protobuf message ride.wallet.v1alpha1.ListTransactionsByBatchIdResponse
 */
export interface ListTransactionsByBatchIdResponse {
    /**
     * @generated from protobuf field: repeated ride.wallet.v1alpha1.Transaction transactions = 1;
     */
    transactions: Transaction[];
}
/**
 * @generated from protobuf message ride.wallet.v1alpha1.ListTransactionsByAccountIdRequest
 */
export interface ListTransactionsByAccountIdRequest {
    /**
     * @generated from protobuf field: string account_id = 1;
     */
    accountId: string;
    /**
     * @generated from protobuf field: optional google.protobuf.FieldMask field_mask = 99;
     */
    fieldMask?: FieldMask;
}
/**
 * @generated from protobuf message ride.wallet.v1alpha1.ListTransactionsByAccountIdResponse
 */
export interface ListTransactionsByAccountIdResponse {
    /**
     * @generated from protobuf field: repeated ride.wallet.v1alpha1.Transaction transactions = 1;
     */
    transactions: Transaction[];
}
// ...
export enum TransactionType {
    /**
     * @generated from protobuf enum value: TRANSACTION_TYPE_UNSPECIFIED = 0;
     */
    UNSPECIFIED = 0,
    /**
     * @generated from protobuf enum value: TRANSACTION_TYPE_DEBIT = 1;
     */
    DEBIT = 1,
    /**
     * @generated from protobuf enum value: TRANSACTION_TYPE_CREDIT = 2;
     */
    CREDIT = 2
}
// @generated message type with reflection information, may provide speed optimized methods
class CreateAccountRequest$Type extends MessageType<CreateAccountRequest> {
    constructor() {
        super("ride.wallet.v1alpha1.CreateAccountRequest", [
            { no: 1, name: "uid", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<CreateAccountRequest>): CreateAccountRequest {
        const message = { uid: "" };
        globalThis.Object.defineProperty(message, MESSAGE_TYPE, { enumerable: false, value: this });
        if (value !== undefined)
            reflectionMergePartial<CreateAccountRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: CreateAccountRequest): CreateAccountRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string uid */ 1:
                    message.uid = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: CreateAccountRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string uid = 1; */
        if (message.uid !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.uid);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message ride.wallet.v1alpha1.CreateAccountRequest
 */
export const CreateAccountRequest = new CreateAccountRequest$Type();
// @generated message type with reflection information, may provide speed optimized methods
class CreateAccountResponse$Type extends MessageType<CreateAccountResponse> {
    constructor() {
        super("ride.wallet.v1alpha1.CreateAccountResponse", [
            { no: 1, name: "account", kind: "message", T: () => Account }
        ]);
    }
    create(value?: PartialMessage<CreateAccountResponse>): CreateAccountResponse {
        const message = {};
        globalThis.Object.defineProperty(message, MESSAGE_TYPE, { enumerable: false, value: this });
        if (value !== undefined)
            reflectionMergePartial<CreateAccountResponse>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: CreateAccountResponse): CreateAccountResponse {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* ride.wallet.v1alpha1.Account account */ 1:
                    message.account = Account.internalBinaryRead(reader, reader.uint32(), options, message.account);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: CreateAccountResponse, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* ride.wallet.v1alpha1.Account account = 1; */
        if (message.account)
            Account.internalBinaryWrite(message.account, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message ride.wallet.v1alpha1.CreateAccountResponse
 */
export const CreateAccountResponse = new CreateAccountResponse$Type();
// ...
export const GetAccountRequest = new GetAccountRequest$Type();
// @generated message type with reflection information, may provide speed optimized methods
class GetAccountResponse$Type extends MessageType<GetAccountResponse> {
    constructor() {
        super("ride.wallet.v1alpha1.GetAccountResponse", [
            { no: 1, name: "account", kind: "message", T: () => Account }
        ]);
    }
    create(value?: PartialMessage<GetAccountResponse>): GetAccountResponse {
        const message = {};
        globalThis.Object.defineProperty(message, MESSAGE_TYPE, { enumerable: false, value: this });
        if (value !== undefined)
            reflectionMergePartial<GetAccountResponse>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: GetAccountResponse): GetAccountResponse {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* ride.wallet.v1alpha1.Account account */ 1:
                    message.account = Account.internalBinaryRead(reader, reader.uint32(), options, message.account);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: GetAccountResponse, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* ride.wallet.v1alpha1.Account account = 1; */
        if (message.account)
            Account.internalBinaryWrite(message.account, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message ride.wallet.v1alpha1.GetAccountResponse
 */
export const GetAccountResponse = new GetAccountResponse$Type();
// ...
export const GetAccountByUidRequest = new GetAccountByUidRequest$Type();
// ...
export const GetAccountByUidResponse = new GetAccountByUidResponse$Type();
// ...
export const Account = new Account$Type();
// ...
export const CreateTransactionsRequest = new CreateTransactionsRequest$Type();
// ...
export const CreateTransactionsRequest_Entry = new CreateTransactionsRequest_Entry$Type();
// ...
export const CreateTransactionsResponse = new CreateTransactionsResponse$Type();
// ...
export const GetTransactionRequest = new GetTransactionRequest$Type();
// ...
export const GetTransactionResponse = new GetTransactionResponse$Type();
// ...
export const ListTransactionsByBatchIdRequest = new ListTransactionsByBatchIdRequest$Type();
// ...
export const ListTransactionsByBatchIdResponse = new ListTransactionsByBatchIdResponse$Type();
// ...
export const ListTransactionsByAccountIdRequest = new ListTransactionsByAccountIdRequest$Type();
// ...
export const ListTransactionsByAccountIdResponse = new ListTransactionsByAccountIdResponse$Type();